pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "A personalized organizational tool with AI integration"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "mcp>=1.27.0",
    "pyyaml>=6.0.3",
    "textual>=8.2.4",
]

[project.scripts]
nexus = "nexus.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
